requirements.txt
# Telegram Bot
python-telegram-bot==20.8

# Битрикс24 Integration
requests==2.31.0
urllib3==2.1.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.1

# Scheduler
APScheduler==3.10.4

# Web Framework (для вебхуков от Битрикс24)
Flask==3.0.0
Flask-CORS==4.0.0

# Utils
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
